backend/requirements.txt
asgiref==3.7.2
Django==4.2.5
django-filter==23.3
drf-extra-fields==3.4.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-colorfield==0.10.1
djoser==2.2.0
gunicorn==20.1.0
Pillow==10.0.1
psycopg2-binary==2.9.8
python-dotenv==1.0.0
pytz==2023.3.post1
reportlab==4.0.5
requests==2.31.0
sqlparse==0.4.4

